requirements.txt
mysql-connector-python~=8.0.28
Flask~=2.1.2